moved isFiltered to utils
