+++ src/utils.js
@@ -94,3 +94,13 @@ export function evalMethod(action, obj) {
   return (new Function('args', `return this.${action.name}(args)`)).apply(obj, args);
 }
 /* eslint-enable */
+
+export function isFiltered(action, localFilter) {
+  if (typeof window === 'undefined' && !localFilter) return true;
+  if (!localFilter) return false;
+  const { whitelist, blacklist } = localFilter;
+  return (
+    whitelist && !action.type.match(whitelist) ||
+    blacklist && action.type.match(blacklist)
+  );
+}
